ipfn_np: Move convergence calculation into the main loop
The total runtime was rougly divided equally between:
 - Summing to calculate the scaling factor
 - Scale the matrix
 - Summing to calculate the convergence

So a 30% runtime reduction can be achieved by reusing the summing in the
first loop to calculate the convergence.  This has the effect that the
convergence reported for each iteration is now rougly equivalent to the
convergence that was previously reported for the iteration before.  So
the new code will run one iteration more to get to the stopping criterion.
(but is also delivers one iteration more accurate results)
This is not a big performance issue, as each iteration is now cheaper,
one extra iteration is not a problem.